markdown: do not use placeholders for escaped characters
This patch recreates the Python Markdown 2.0.3 behavior where escaped
characters are simply unescaped. Python Markdown 2.1 replaces them with
placeholders which makes it impossible to decide if it is some inline
HTML, some escaped character or anything else. Additionally it is not
possible to apply special treatment to the escaped characters because
they are hidden.